move source to /src and update tests to reflect new directory stucture
