+++ javascript_client/__tests__/cliTest.js
@@ -3,11 +3,11 @@ var childProcess = require("child_process")
 describe("CLI", () => {
   it("exits 1 on error", () => {
     expect(() => {
-      childProcess.execSync("./cli.js sync", {stdio: "pipe"})
+      childProcess.execSync("./src/cli.js sync", {stdio: "pipe"})
     }).toThrow("URL must be provided for sync")
   })
 
   it("exits 0 on OK", () => {
-    childProcess.execSync("./cli.js sync -h", {stdio: "pipe"})
+    childProcess.execSync("./src/cli.js sync -h", {stdio: "pipe"})
   })
 })
