+++ javascript_client/__tests__/indexTest.js
@@ -1,4 +1,4 @@
-var graphqlRubyClient = require("../index.js")
+var graphqlRubyClient = require("../src/index.js")
 
 it("exports the sync function", () => {
   expect(graphqlRubyClient.sync).toBeInstanceOf(Function)
