+++ javascript_client/src/subscriptions/__tests__/addGraphQLSubscriptionsTest.js
