+++ javascript_client/src/subscriptions/__tests__/registryTest.js
