+++ javascript_client/src/subscriptions/addGraphQLSubscriptions.js
