+++ javascript_client/src/subscriptions/registry.js
