+++ javascript_client/src/sync/__tests__/generateClientTest.js
