+++ javascript_client/src/sync/__tests__/generateJsonClientTest.js
