+++ javascript_client/src/sync/__tests__/prepareProjectTest.js
