+++ javascript_client/src/sync/__tests__/sendPayloadTest.js
