+++ javascript_client/src/sync/addTypenameToSelectionSet.js
