+++ javascript_client/src/sync/generateClient.js
