+++ javascript_client/src/sync/logger.js
