+++ javascript_client/src/sync/prepareIsolatedFiles.js
