+++ javascript_client/src/sync/prepareProject.js
