+++ javascript_client/src/sync/sendPayload.js
